Bug 2027687 - Natively serialize webrtc::VideoType in camera::VideoCaptureCapability. r=tschuster,webrtc-reviewers,bwc
Differential Revision: https://phabricator.services.mozilla.com/D296673

UltraBlame original commit: d3b541449932676f8a506aad07696d5a3656dba2

+++ dom/media/systemservices/CamerasParent.cpp
@@ -1086,7 +1086,7 @@ ipc::IPCResult CamerasParent::RecvGetCaptureCapability(
             auto webrtcCaps = aValue.ResolveValue();
             VideoCaptureCapability capCap(
                 webrtcCaps.width, webrtcCaps.height, webrtcCaps.maxFPS,
-                static_cast<int>(webrtcCaps.videoType), webrtcCaps.interlaced);
+                webrtcCaps.videoType, webrtcCaps.interlaced);
             LOG("Capability: %u %u %u %d %d", webrtcCaps.width,
                 webrtcCaps.height, webrtcCaps.maxFPS,
                 static_cast<int>(webrtcCaps.videoType), webrtcCaps.interlaced);
